use https for gravatar url if site uses it
If avatar uses http to get gravatar-images, browsers will complain when
retrospectiva runs in a secured (https) environment. This patch adds an
option to the gravatar method and it's default value depends on the
site_url setting. If retrospectiva runs within a https server, it will
use secure connection, non-secure connection otherwise.